+++ unix/src/machine_stats/__main__.py
@@ -1,0 +1,6 @@
+"""
+Script to prepare JSON output for tidal sync servers from the list of hosts
+"""
+from machine_stats import main
+
+main()
